Load recipes into value objects on startup
