+++ src/main/java/com/wd40/satisfactorymanager/data/Ingredient.java
@@ -1,0 +1,10 @@
+package com.wd40.satisfactorymanager.data;
+
+import lombok.Value;
+
+@Value
+public class Ingredient {
+
+  String item;
+  int amount;
+}
