+++ src/main/java/com/wd40/satisfactorymanager/data/Recipe.java
@@ -1,0 +1,17 @@
+package com.wd40.satisfactorymanager.data;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.Set;
+import lombok.Value;
+
+@Value
+public class Recipe {
+
+  @JsonProperty("key")
+  String keySegment;
+
+  String name;
+  String machineType;
+  Set<Ingredient> inputs;
+  Set<Ingredient> outputs;
+}
