--- src/main/java/com/wd40/satisfactorymanager/model/Ingredient.java
@@ -1,41 +1,0 @@
-package com.wd40.satisfactorymanager.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class Ingredient {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Integer id;
-
-  private String item;
-  private int amount;
-
-  public Ingredient() {}
-
-  @JsonIgnore
-  public Integer getId() {
-    return id;
-  }
-
-  public String getItem() {
-    return item;
-  }
-
-  public void setItem(String item) {
-    this.item = item;
-  }
-
-  public int getAmount() {
-    return amount;
-  }
-
-  public void setAmount(int amount) {
-    this.amount = amount;
-  }
-}
